Renames icon file names
