+++ packages/mobile-top-nav/src/assets/img/icon-audio.js
