+++ packages/mobile-top-nav/src/assets/img/icon-software.js
