+++ packages/mobile-top-nav/src/assets/img/icons.js
@@ -1,7 +1,7 @@
-import './icon_audio';
-import './icon_ellipses';
-import './icon_images';
-import './icon_software';
-import './icon_texts';
-import './icon_video';
-import './icon_web';
+import './icon-audio';
+import './icon-ellipses';
+import './icon-images';
+import './icon-software';
+import './icon-texts';
+import './icon-video';
+import './icon-web';
